clarify what doesn't work as a baseURL
I was confused when reading the sample config file, as explained in
issue #164. "don't support /" could have meant a trailing slash, or
subdirectories, or at least wasn't exactly clear to me. By adding an
explicit example of a broken behavior, we make it clear what we're
talking about here, and what to avoid.

Closes: #164