Fix relative date for resolved issues and non-UTC browser timezone.
The missing <UTC> in the layout causes a weird glitch if the users browser
is in a non-UTC timezone. For open incidents, the time is
rendered as "23-07-2021 12:17:40 UTC", and then the javascript library
correctly translates from UTC into the users browser time, resulting
in say "30m ago".

For a resolved incident, the timezone of the date was stripped due
to the missing <UTC> in the format string, resulting in "23-07-2021 12:17:40".
If the user has a browser is in the timezone CEST (UTC+2), this causes
the relative time to jump from "30 minutes ago" for an open incident
to "2 hours ago" for a closed incident, because now the time without
time zone is not interpreted as UTC anymore.