layout: use simple arrows as list pagination icons
The currently used "triangle-headed" arrows are from the Unicode code
block "Supplemental Arrows-C" [1], which seems to be not widely
supported. This also applies to e.g. Roboto, which is being explicitly
used in the CSS.

Instead, use the "simple" arrows from the "Arrows" block, which is much
more widely supported by fonts.

[1]: https://en.wikipedia.org/wiki/Supplemental_Arrows-C

closes #242